requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
sqlalchemy==2.0.25
alembic==1.13.1
aiosqlite==0.19.0
httpx==0.26.0
pyyaml==6.0.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
sse-starlette==1.8.2
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
